+++ src/dotnet/UI.Blazor.App/Components/ChatMessageEditor/IAttachmentFilePicker.cs
@@ -2,7 +2,7 @@ using ActualChat.UI.Blazor.App.Services;
 
 namespace ActualChat.UI.Blazor.App.Components;
 
-public record AttachFileInfo(string FileName, string FileType, int Length, IFileProvider FileProvider);
+public record AttachFileInfo(string FileName, string FileType, long Length, IFileProvider FileProvider);
 
 public interface IAttachmentFilePicker
 {
